chore: move QuoteResults to types
